requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.7
django-environ==0.11.2
django-filter==24.2
django-guardian==2.4.0
djangorestframework==3.15.2
idna==3.10
Markdown==3.6
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
PyYAML==6.0.1